Fix sled store schema migrations (#121)

+++ src/store/sled.rs
@@ -22,7 +22,7 @@ use libsignal_service::{
     push_service::DEFAULT_DEVICE_ID,
     ServiceAddress,
 };
-use log::{debug, error, info, trace, warn};
+use log::{debug, error, trace, warn};
 use matrix_sdk_store_encryption::StoreCipher;
 use prost::Message;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
@@ -58,13 +58,14 @@ pub struct SledStore {
 
 /// Sometimes Migrations can't proceed without having to drop existing
 /// data. This allows you to configure, how these cases should be handled.
-#[derive(PartialEq, Eq, Clone, Debug)]
+#[derive(Default, PartialEq, Eq, Clone, Debug)]
 pub enum MigrationConflictStrategy {
     /// Just drop the data, we don't care that we have to register or link again
     Drop,
     /// Raise a `Error::MigrationConflict` error with the path to the
     /// DB in question. The caller then has to take care about what they want
     /// to do and try again after.
+    #[default]
     Raise,
     /// _Default_: The _entire_ database is backed up under, before the databases are dropped.
     BackupAndDrop,
@@ -239,38 +240,33 @@ fn migrate(
         for step in schema_version.steps() {
             match &step {
                 SchemaVersion::V1 => {
-                    warn!("migrating from v0, nothing to do")
+                    debug!("migrating from v0, nothing to do")
                 }
                 SchemaVersion::V2 => {
-                    info!("migrating from schema v1 to v2: encrypting state if cipher is enabled");
+                    debug!("migrating from schema v1 to v2: encrypting state if cipher is enabled");
 
                     // load registration data the old school way
-                    let data = store
-                        .db
-                        .get(SLED_KEY_REGISTRATION)?
-                        .ok_or(Error::NotYetRegisteredError)?;
-                    let state = serde_json::from_slice(&data).map_err(Error::from)?;
+                    if let Some(data) = store.db.get(SLED_KEY_REGISTRATION)? {
+                        let state = serde_json::from_slice(&data).map_err(Error::from)?;
 
-                    // save it the new school way
-                    store.save_state(&state)?;
+                        // save it the new school way
+                        store.save_state(&state)?;
 
-                    // remove old data
-                    store.db.remove(SLED_KEY_REGISTRATION)?;
+                        // remove old data
+                        store.db.remove(SLED_KEY_REGISTRATION)?;
+                    }
                 }
                 _ => return Err(Error::MigrationConflict),
             }
 
             store.insert(SLED_TREE_STATE, SLED_KEY_SCHEMA_VERSION, step)?;
+            store.db.flush()?;
         }
 
         Ok(())
     };
 
-    let migration_res = run_migrations();
-
-    if let Err(Error::MigrationConflict) = migration_res {
-        let db = sled::open(db_path)?;
-
+    if let Err(error) = run_migrations() {
         match migration_conflict_strategy {
             MigrationConflictStrategy::BackupAndDrop => {
                 let mut new_db_path = db_path.to_path_buf();
@@ -283,17 +279,12 @@ fn migrate(
                 ));
                 fs_extra::dir::create_all(&new_db_path, false)?;
                 fs_extra::dir::copy(db_path, new_db_path, &fs_extra::dir::CopyOptions::new())?;
-
-                for tree in db.tree_names() {
-                    db.drop_tree(tree)?;
-                }
+                fs_extra::dir::remove(db_path)?;
             }
             MigrationConflictStrategy::Drop => {
-                for tree in db.tree_names() {
-                    db.drop_tree(tree)?;
-                }
+                fs_extra::dir::remove(db_path)?;
             }
-            MigrationConflictStrategy::Raise => migration_res?,
+            MigrationConflictStrategy::Raise => return Err(error),
         }
     }
